Added documentation
