+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/ActionFunctions/SpecimenSide/MoveArmToScoringSpecimenPosition.java
@@ -5,7 +5,7 @@ import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 
 import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.ActionFunction;
-import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.BlackBoardSingleton;
+import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.BlackBoard;
 import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.Status;
 import org.firstinspires.ftc.teamcode.hardwareControl.actuators.arm.ArmController;
 
@@ -33,7 +33,7 @@ public class MoveArmToScoringSpecimenPosition implements ActionFunction
         telemetry.clearAll();
     }
 
-    public Status perform(BlackBoardSingleton blackBoard) {
+    public Status perform(BlackBoard blackBoard) {
         Status status;
 
         if (lastStatus == Status.SUCCESS) {
