+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/ActionFunctions/SpecimenSide/ScoreSpecimen1.java
@@ -8,7 +8,7 @@ import com.pedropathing.pathgen.Point;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 
 import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.ActionFunction;
-import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.BlackBoardSingleton;
+import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.BlackBoard;
 import org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general.Status;
 import org.firstinspires.ftc.teamcode.hardwareControl.actuators.driveTrain.DriveTrainController;
 
@@ -37,7 +37,7 @@ public class ScoreSpecimen1 implements ActionFunction
                 .build();
     }
 
-    public Status perform(BlackBoardSingleton blackBoard) {
+    public Status perform(BlackBoard blackBoard) {
         Status status;
 
         if (lastStatus == Status.SUCCESS) {
