+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/general/ActionFunction.java
@@ -1,7 +1,5 @@
 package org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general;
 
 public interface ActionFunction {
-    public Status perform(BlackBoardSingleton blackBoard);
-
-    Status perform(BlackBoardSingleton blackBoard);
+    Status perform(BlackBoard blackBoard);
 }
