+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/general/Condition.java
@@ -1,5 +1,5 @@
 package org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general;
 
 public interface Condition {
-    boolean check(BlackBoardSingleton blackBoard);
+    boolean check(BlackBoard blackBoard);
 }
