+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/general/Node.java
@@ -1,5 +1,5 @@
 package org.firstinspires.ftc.teamcode.behaviorControl.BehaviorTrees.BehaviorTreeComponents.general;
 
 public abstract class Node {
-    public abstract Status execute(BlackBoardSingleton globalStore);
+    public abstract Status execute(BlackBoard globalStore);
 }
